+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## 2.3.0
 
 - Fix tooltip.
+- Fix image size when single fill enable
 - Allow using variable for `base URL`, `Suffix`, `Image width` and `Image height`.
 - Now when something is wrong an error is thrown instead of displaying a div. This will cause grafana to display the
   error in the top left corner of the panel.
-- When an image fail to load, a warning is logged containing the
-  url see [#11](https://github.com/Dalvany/dalvany-image-panel/issues/11).
+- When an image fail to load, a warning is logged containing the url
+  see [#11](https://github.com/Dalvany/dalvany-image-panel/issues/11).
+- Possibility to add a square as overlay over pictures. Color can be chosen with a mapping for field values.
 
 ## 2.2.0